Move windowsRegistryLocator.ts
